Move existing fixed time window lesson into windowing section
