+++ learning/katas/go/windowing/fixed_time_window/fixed_time_window/cmd/main.go
@@ -16,8 +16,8 @@
 package main
 
 import (
-	"beam.apache.org/learning/katas/core_transforms/windowing/windowing/pkg/common"
-	"beam.apache.org/learning/katas/core_transforms/windowing/windowing/pkg/task"
+	"beam.apache.org/learning/katas/windowing/fixed_time_window/fixed_time_window/pkg/common"
+	"beam.apache.org/learning/katas/windowing/fixed_time_window/fixed_time_window/pkg/task"
 	"context"
 	"github.com/apache/beam/sdks/go/pkg/beam"
 	"github.com/apache/beam/sdks/go/pkg/beam/log"
