+++ learning/katas/go/windowing/fixed_time_window/fixed_time_window/pkg/task/task.go
