+++ learning/katas/go/windowing/fixed_time_window/fixed_time_window/task-remote-info.yaml
