+++ learning/katas/go/windowing/fixed_time_window/lesson-remote-info.yaml
